backend/requirements.txt
# Web framework
# Unpinned so pip picks the latest versions compatible with your Python
fastapi>=0.115.0
uvicorn>=0.30.0          # No [standard] extra — avoids the greenlet/uvloop build issue on Python 3.14
python-multipart>=0.0.12

# Resume parsing
PyPDF2>=3.0.1
python-docx>=1.1.0

# AI
anthropic>=0.27.0

# Resume generation
reportlab>=4.2.0
fpdf2>=2.8.0

# HTTP / scraping
httpx>=0.27.0
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.2.0

# Browser automation (optional — install separately if needed)
# playwright>=1.48.0

# Utilities
pydantic>=2.9.0          # 2.9+ ships pre-built wheels for Python 3.14
python-dotenv>=1.0.1
aiofiles>=24.1.0
tenacity>=8.3.0
